+++ src/component/Laptops-list/laptops-list.jsx
